split skydb v2 tests into two files to improve performance
